Tree and stem tables from Barro Colorado Island (bci).
Source: R/data-census.R

For a general description see census_description. These datasets are tree and stem tables of the same trees selected to create the tables bci_vft_1ha and bci_vft_random. In the name of the tree and stem tables documented here, the number after stem or census matches a value of the column PlotCensusNumber in the corresponding ViewFullTable.

bci_stem6_1ha

bci_stem7_1ha

bci_tree6_1ha

bci_tree7_1ha

bci_stem6_random

bci_stem7_random

bci_tree6_random

bci_tree7_random

Format

See examples.

See also

bci_vft_1ha, bci_vft_random, census_description.

Examples

str(bci_stem6_1ha, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	10373 obs. of  16 variables:
#>  $ treeID   : num  1853 1857 1861 1864 1866 ...
#>  $ stemID   : num  1853 1857 1861 1864 1866 ...
#>  $ tag      : chr  "001854" "001858" "001862" "001865" ...
#>  $ StemTag  : chr  "" "" "" "" ...
#>  $ sp       : chr  "simaam" "poutre" "quaras" "guargr" ...
#>  $ quadrat  : chr  "3909" "3909" "3908" "3908" ...
#>  $ gx       : num  792 793 783 799 793 ...
#>  $ gy       : num  181 190 177 168 170 ...
#>  $ dbh      : num  508 415 774 300 450 576 226 500 314 583 ...
#>  $ hom      : num  1.3 3.7 6 1.3 3.75 6.8 3 3.8 1.3 1.3 ...
#>  $ ExactDate: chr  "2005-07-18" "2005-09-30" "2005-10-12" "2005-07-14" ...
#>  $ DFstatus : chr  "alive" "alive" "alive" "alive" ...
#>  $ codes    : chr  NA "B;cylN" "B;cylN" NA ...
#>  $ date     : num  16635 16709 16721 16631 16708 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ agb      : num  1.761 2.193 5.749 0.658 2.285 ...

str(bci_tree7_1ha, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	4822 obs. of  16 variables:
#>  $ treeID   : num  1853 1857 1861 1864 1866 ...
#>  $ tag      : chr  "001854" "001858" "001862" "001865" ...
#>  $ sp       : chr  "simaam" "poutre" "quaras" "guargr" ...
#>  $ quadrat  : chr  "3909" "3909" "3908" "3908" ...
#>  $ gx       : num  792 793 783 799 793 ...
#>  $ gy       : num  181 190 177 168 170 ...
#>  $ stemID   : int  1853 1857 1861 1864 1866 1868 1869 1870 1871 1872 ...
#>  $ dbh      : num  555 455 778 304 454 595 229 500 335 560 ...
#>  $ hom      : num  1.3 3.7 6 1.3 3.75 6.8 3 3.8 1.3 3 ...
#>  $ ExactDate: chr  "2010-07-27" "2010-09-07" "2010-09-10" "2010-07-23" ...
#>  $ codes    : chr  NA "B;cylN" "B;cylN" NA ...
#>  $ date     : num  18470 18512 18515 18466 18512 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ nostems  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ agb      : num  2.189 2.763 5.819 0.681 2.336 ...
#>  $ ba       : num  0.2419 0.1626 0.4754 0.0726 0.1619 ...

str(bci_stem7_random, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2266 obs. of  16 variables:
#>  $ treeID   : num  590 1057 1451 2362 2867 ...
#>  $ stemID   : num  590 1057 1451 2362 2867 ...
#>  $ tag      : chr  "000580" "001054" "001449" "002369" ...
#>  $ StemTag  : chr  "" "" "" "" ...
#>  $ sp       : chr  "virosp" "alsebl" "quaras" "quaras" ...
#>  $ quadrat  : chr  "4614" "4317" "4113" "3618" ...
#>  $ gx       : num  933 866 831 737 669 ...
#>  $ gy       : num  283 358 277 371 194 ...
#>  $ dbh      : num  304 365 715 645 433 378 240 860 242 605 ...
#>  $ hom      : num  4 3 6.35 5.85 3 3.6 3.7 1.3 1.3 4 ...
#>  $ ExactDate: chr  "2010-09-21" "2010-09-09" "2010-09-13" "2010-09-10" ...
#>  $ DFstatus : chr  "alive" "alive" "alive" "alive" ...
#>  $ codes    : chr  "B;cylY" "B;cylN" "B;cylN" "B;cylN" ...
#>  $ date     : num  18526 18514 18518 18515 18506 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ agb      : num  0.572 1.07 4.765 3.723 1.613 ...

str(bci_tree6_random, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1000 obs. of  16 variables:
#>  $ treeID   : num  590 1057 1451 2362 2867 ...
#>  $ tag      : chr  "000580" "001054" "001449" "002369" ...
#>  $ sp       : chr  "virosp" "alsebl" "quaras" "quaras" ...
#>  $ quadrat  : chr  "4614" "4317" "4113" "3618" ...
#>  $ gx       : num  933 866 831 737 669 ...
#>  $ gy       : num  283 358 277 371 194 ...
#>  $ stemID   : int  590 1057 1451 2362 2867 2912 3477 3492 3707 4822 ...
#>  $ dbh      : num  330 344 857 640 411 371 233 823 238 585 ...
#>  $ hom      : num  1.3 3 4 5.85 3 3.6 3.7 1.3 1.3 4 ...
#>  $ ExactDate: chr  "2005-09-19" "2005-10-12" "2005-10-11" "2005-10-11" ...
#>  $ codes    : chr  NA "B;cylN" "B;cylN" "B;cylN" ...
#>  $ date     : num  16698 16721 16720 16720 16711 ...
#>  $ status   : chr  "A" "A" "A" "A" ...
#>  $ nostems  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ agb      : num  0.706 0.919 7.294 3.654 1.414 ...
#>  $ ba       : num  0.0855 0.0929 0.5768 0.3217 0.1327 ...