All threes in one hectare from Barro Colorado Island..
Source: R/data.R

One hectare of data for hectare: px >= 700 and px < 800; and py >=100 and py < 200. The data includes only those records from census 2005 and census 2010.

vft_1ha_bci

Format

An object of class tbl_df (inherits from tbl, data.frame) with 10543 rows and 32 columns.

Details

This is a slightly cleaned version from a ViewFullTable table. For a raw .csv version see http://bit.ly/fgeo-opendata-vft-1ha-bci.

Examples

str(vft_1ha_bci, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	10543 obs. of  32 variables:
#>  $ DBHID           : int  148842 150985 148877 148912 151013 149016 151095 149050 151122 149083 ...
#>  $ PlotName        : chr  "bci" "bci" "bci" "bci" ...
#>  $ PlotID          : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Family          : chr  "Simaroubaceae" "Simaroubaceae" "Sapotaceae" "Sapotaceae" ...
#>  $ Genus           : chr  "Simarouba" "Simarouba" "Pouteria" "Pouteria" ...
#>  $ SpeciesName     : chr  "amara" "amara" "reticulata" "reticulata" ...
#>  $ Mnemonic        : chr  "simaam" "simaam" "poutre" "poutre" ...
#>  $ Subspecies      : chr  NA NA NA NA ...
#>  $ SpeciesID       : int  935 935 1155 1155 1155 871 871 434 434 326 ...
#>  $ SubspeciesID    : chr  NA NA NA NA ...
#>  $ QuadratName     : chr  "3909" "3909" "3909" "3909" ...
#>  $ QuadratID       : int  2806 2806 2806 2806 2806 2805 2805 2805 2805 2805 ...
#>  $ PX              : num  792 792 793 793 793 ...
#>  $ PY              : num  181 181 190 190 190 ...
#>  $ QX              : num  12.1 12.1 12.8 12.8 12.8 3.2 3.2 19.1 19.1 13.4 ...
#>  $ QY              : num  1.2 1.2 9.6 9.6 9.6 16.8 16.8 7.8 7.8 10.5 ...
#>  $ TreeID          : int  1853 1853 1857 1857 1857 1861 1861 1864 1864 1866 ...
#>  $ Tag             : chr  "001854" "001854" "001858" "001858" ...
#>  $ StemID          : int  1853 1853 1857 1857 1857 1861 1861 1864 1864 1866 ...
#>  $ StemNumber      : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ StemTag         : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ PrimaryStem     : chr  "main" "main" "main" "main" ...
#>  $ CensusID        : int  6 171 6 6 171 6 171 6 171 6 ...
#>  $ PlotCensusNumber: int  6 7 6 6 7 6 7 6 7 6 ...
#>  $ DBH             : num  508 555 494 415 455 774 778 300 304 496 ...
#>  $ HOM             : num  1.3 1.3 1.3 3.7 3.7 6 6 1.3 1.3 1.3 ...
#>  $ ExactDate       : Date, format: "2005-07-18" "2010-07-27" ...
#>  $ Date            : int  16635 18470 16709 16709 18512 16721 18515 16631 18466 16708 ...
#>  $ ListOfTSM       : chr  NA NA "B;cylN" "B;cylN" ...
#>  $ HighHOM         : int  1 1 0 1 1 1 1 1 1 0 ...
#>  $ LargeStem       : chr  NA NA NA NA ...
#>  $ Status          : chr  "alive" "alive" "alive" "alive" ...