One Thousand trees selected at random from Barro Colorado Island.
Source: R/data.R

One Thousand, randomly chosen, trees from Barro Colorado Island (from the whole plot). Only included data from census 2005 and census 2010.

vft_random_bci

Format

An object of class tbl_df (inherits from tbl, data.frame) with 2257 rows and 32 columns.

Details

This is a slightly cleaned version from a ViewFullTable table. For a raw .csv version see http://bit.ly/fgeo-opendata-vft-random-bci.

Examples

str(vft_random_bci, give.attr = FALSE)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2257 obs. of  32 variables:
#>  $ DBHID           : int  1621522 1619521 1632477 1630644 1077531 1074697 826469 823306 1006845 1003843 ...
#>  $ PlotName        : chr  "bci" "bci" "bci" "bci" ...
#>  $ PlotID          : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Family          : chr  "Violaceae" "Violaceae" "Fabaceae-papilionoideae" "Fabaceae-papilionoideae" ...
#>  $ Genus           : chr  "Hybanthus" "Hybanthus" "Lonchocarpus" "Lonchocarpus" ...
#>  $ SpeciesName     : chr  "prunifolius" "prunifolius" "heptaphyllus" "heptaphyllus" ...
#>  $ Mnemonic        : chr  "hybapr" "hybapr" "loncla" "loncla" ...
#>  $ Subspecies      : chr  NA NA NA NA ...
#>  $ SpeciesID       : int  494 494 1134 1134 99 99 895 895 980 980 ...
#>  $ SubspeciesID    : chr  NA NA NA NA ...
#>  $ QuadratName     : chr  "4715" "4715" "0022" "0022" ...
#>  $ QuadratID       : int  3012 3012 1261 1261 2241 2241 3052 3052 2315 2315 ...
#>  $ PX              : num  957.3 957.3 16.3 16.3 324.6 ...
#>  $ PY              : num  307 307 442 442 392 ...
#>  $ QX              : num  17.3 17.3 16.3 16.3 4.6 4.6 0.3 0.3 0.6 0.6 ...
#>  $ QY              : num  6.9 6.9 2.5 2.5 12.4 12.4 18.2 18.2 1 1 ...
#>  $ TreeID          : int  267125 267125 269374 269374 159260 159260 109920 109920 145725 145725 ...
#>  $ Tag             : chr  "312412" "312412" "400860" "400860" ...
#>  $ StemID          : int  267125 267125 577077 577077 159260 159260 109920 109920 145725 145725 ...
#>  $ StemNumber      : int  1 1 2 2 1 1 1 1 1 1 ...
#>  $ StemTag         : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ PrimaryStem     : chr  "main" "main" "main" "main" ...
#>  $ CensusID        : int  6 171 6 171 6 171 6 171 6 171 ...
#>  $ PlotCensusNumber: int  6 7 6 7 6 7 6 7 6 7 ...
#>  $ DBH             : num  27 30 NA NA 42 NA 26 28 28 40 ...
#>  $ HOM             : num  1.3 1.3 NA NA 1.3 NA 1.3 1.3 1.3 1.3 ...
#>  $ ExactDate       : Date, format: "2005-10-05" "2010-09-06" ...
#>  $ Date            : int  16714 18511 16492 18323 16544 18378 16702 18507 16572 18395 ...
#>  $ ListOfTSM       : chr  NA NA "OR" "D;N" ...
#>  $ HighHOM         : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ LargeStem       : chr  NA NA NA NA ...
#>  $ Status          : chr  "alive" "alive" "broken below" "dead" ...

show_some_values_of_each_column <- function(x) head(unique(x))
lapply(vft_random_bci, show_some_values_of_each_column)

#> $DBHID
#> [1] 1621522 1619521 1632477 1630644 1077531 1074697
#> 
#> $PlotName
#> [1] "bci"
#> 
#> $PlotID
#> [1] 1
#> 
#> $Family
#> [1] "Violaceae"               "Fabaceae-papilionoideae"
#> [3] "Lauraceae"               "Apocynaceae"            
#> [5] "Sapindaceae"             "Burseraceae"            
#> 
#> $Genus
#> [1] "Hybanthus"     "Lonchocarpus"  "Beilschmiedia" "Rinorea"      
#> [5] "Swartzia"      "Aspidosperma" 
#> 
#> $SpeciesName
#> [1] "prunifolius"  "heptaphyllus" "pendula"      "sylvatica"    "simplex"     
#> [6] "spruceanum"  
#> 
#> $Mnemonic
#> [1] "hybapr" "loncla" "beilpe" "rinosy" "swars2" "aspicr"
#> 
#> $Subspecies
#> [1] NA            "ochnacea"    "grandiflora"
#> 
#> $SpeciesID
#> [1]  494 1134   99  895  980 1112
#> 
#> $SubspeciesID
#> [1] NA   "8"  "13" "7"  "15"
#> 
#> $QuadratName
#> [1] "4715" "0022" "1619" "4905" "1918" "2320"
#> 
#> $QuadratID
#> [1] 3012 1261 2241 3052 2315 2417
#> 
#> $PX
#> [1] 957.30  16.30 324.60 980.30 380.60 464.76
#> 
#> $PY
#> [1] 306.90 442.50 392.40 118.20 361.00 403.78
#> 
#> $QX
#> [1] 17.30 16.30  4.60  0.30  0.60  4.76
#> 
#> $QY
#> [1]  6.90  2.50 12.40 18.20  1.00  3.78
#> 
#> $TreeID
#> [1] 267125 269374 159260 109920 145725 468654
#> 
#> $Tag
#> [1] "312412" "400860" "183595" "125801" "167467" "823295"
#> 
#> $StemID
#> [1] 267125 577077 159260 109920 145725 468227
#> 
#> $StemNumber
#> [1]  1  2 19 20 21 22
#> 
#> $StemTag
#> [1] NA  1  2  3  4  5
#> 
#> $PrimaryStem
#> [1] "main"      "secondary" NA         
#> 
#> $CensusID
#> [1]   6 171
#> 
#> $PlotCensusNumber
#> [1] 6 7
#> 
#> $DBH
#> [1] 27 30 NA 42 26 28
#> 
#> $HOM
#> [1] 1.30   NA 1.20 4.00 3.50 5.85
#> 
#> $ExactDate
#> [1] "2005-10-05" "2010-09-06" "2005-02-25" "2010-03-02" "2005-04-18"
#> [6] "2010-04-26"
#> 
#> $Date
#> [1] 16714 18511 16492 18323 16544 18378
#> 
#> $ListOfTSM
#> [1] NA    "OR"  "D;N" "D;C" "L"   "M"  
#> 
#> $HighHOM
#> [1] 1 0
#> 
#> $LargeStem
#> [1] NA
#> 
#> $Status
#> [1] "alive"        "broken below" "dead"         "missing"     
#>